pyproject.toml
[tool.poetry]
name = "agentic-system-from-scratch"
version = "0.1.0"
description = ""
authors = ["yenslife <77geo5rge6@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
rich = "^13.9.4"
openai = "^1.60.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
